Some cleanup.

+++ work_tree.cpp
@@ -1,17 +1,12 @@
-#include <memory>
-#include <initializer_list>
-#include <iostream>
-#include <cassert>
-#include <map>
-#include <iterator>
-#include <algorithm>
-#include <limits>
-
 #include "work_tree.hpp"
 
-int work_tree_node::id_counter = 0;
+#include <iostream>
+
+size_t work_tree_node::id_counter = 0;
 
 int main(int argc, char const *argv[]) {
+
+
 	work_tree asd({1, 2, 4, 2});
 
 	std::cout << "=== COMPACT ===" << std::endl;
